pyproject.toml
[project]
name = "maigret-mcp-server"
version = "2.0.0"
description = "AI-enabled OSINT pipeline: Maigret → PPF → Kafka, exposed as MCP tools and FastAPI endpoints."
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "confluent-kafka>=2.3.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "mcp>=1.0.0",
    "groq>=0.5.0",
    "httpx>=0.27.0",
    "splink>=4.0.0",
    "pandas>=2.0.0",
    "pypdf>=4.2.0",
    "python-docx>=1.1.2",
    "holehe>=1.61",
    "h8mail>=2.0.0",
    "octosuite>=2.0.0",
    "linkedin_scraper>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[project.scripts]
maigret-server = "maigret.main:run"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["maigret/tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.setuptools.packages.find]
include = ["maigret*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
